+++ vllm/executor/ray_distributed_executor.py
@@ -25,7 +25,7 @@ from vllm.utils import (
     get_ip,
     get_open_port,
 )
-from vllm.utils.func import make_async
+from vllm.utils.async_utils import make_async
 from vllm.v1.outputs import SamplerOutput
 
 if ray is not None:
